Decreased bot.py's complexity
